the ladder finds its rungs: the fingering takes its numbered seat among the scripts, and the bow that trails a fermata becomes a mover itself, one tenth above

+++ LilySharp.Core/Svg/Layout/ArticulationSpacing.cs
@@ -91,7 +91,7 @@ internal static class ArticulationSpacing
     /// plain rest (<c>r4@fermata</c>, fixture feature-tour) IS an ordinary Script.
     /// </para>
     /// </remarks>
-    public static int? OutsideStaffPriority(ArticulationType type) => type switch
+    public static double? OutsideStaffPriority(ArticulationType type) => type switch
     {
         ArticulationType.Fermata or ArticulationType.FermataShort
             or ArticulationType.FermataLong => 75,
